+++ interpret/interpret_test.py
@@ -1,5 +1,6 @@
 from interpret import interpret
 from interpret.test import test_environment
+from numpy import int32
 from stack_machine import ExpressionStack
 from stack_machine import NumberValue
 from syntax_tree import Block
@@ -14,9 +15,9 @@ def test_interpret():
             ExpressionStatement(
                 Expression(
                     [
-                        NumberTerm(1),
-                        NumberTerm(2),
-                        NumberTerm(3),
+                        NumberTerm(int32(1)),
+                        NumberTerm(int32(2)),
+                        NumberTerm(int32(3)),
                     ],
                     comment=None,
                     block=None,
@@ -28,5 +29,9 @@ def test_interpret():
     environment = test_environment()
     interpret(block, expression_stack, environment)
     assert expression_stack == ExpressionStack(
-        [NumberValue(1), NumberValue(2), NumberValue(3)]
+        [
+            NumberValue(int32(1)),
+            NumberValue(int32(2)),
+            NumberValue(int32(3)),
+        ]
     )
